chore: load opentelemetry modules on demand
