+++ apps/mail-bridge/tracing.ts
@@ -1,4 +1,7 @@
-import { setupOpentelemetry } from '@u22n/otel';
+import { opentelemetryEnabled } from '@u22n/otel';
 import { name, version } from './package.json';
 
-setupOpentelemetry({ name, version });
+if (opentelemetryEnabled) {
+  const { setupOpentelemetry } = await import('@u22n/otel/setup');
+  setupOpentelemetry({ name, version });
+}
