+++ apps/web/src/instrumentation.ts
@@ -1,7 +1,11 @@
 export async function register() {
   if (process.env.NEXT_RUNTIME === 'nodejs') {
-    const { name, version } = await import('../package.json');
-    const { setupOpentelemetry } = await import('@u22n/otel');
-    setupOpentelemetry({ name, version });
+    const { opentelemetryEnabled } = await import('@u22n/otel');
+
+    if (opentelemetryEnabled) {
+      const { name, version } = await import('../package.json');
+      const { setupOpentelemetry } = await import('@u22n/otel/setup');
+      setupOpentelemetry({ name, version });
+    }
   }
 }
