+++ packages/otel/exports.ts
@@ -1,2 +1,1 @@
-export { trace } from '@opentelemetry/api';
 export { flatten, unflatten } from 'flat';
