+++ packages/otel/package.json
@@ -4,6 +4,7 @@
   "type": "module",
   "exports": {
     ".": "./index.ts",
+    "./setup": "./setup.ts",
     "./hono": "./hono.ts",
     "./exports": "./exports.ts",
     "./logger": "./logger.ts",
